fastapi/requirements.txt
fastapi==0.78.0
python-dotenv==0.20.0
requests==2.27.1
beautifulsoup4==4.10.0
uvicorn==0.17.6
